# By Todd Gamblin
# Via Todd Gamblin
* commit 'b4fddad7eff448adf701fc9e88cf02cd6e582f15':
  Fix for SPACK-10: Spack now forks before install()

+++ lib/spack/spack/directory_layout.py
@@ -86,8 +86,11 @@ class DirectoryLayout(object):
             shutil.rmtree(path, True)
 
         path = os.path.dirname(path)
-        while not os.listdir(path) and path != self.root:
-            os.rmdir(path)
+        while path != self.root:
+            if os.path.isdir(path):
+                if os.listdir(path):
+                    return
+                os.rmdir(path)
             path = os.path.dirname(path)
